% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{contourStewart}
\alias{contourStewart}
\title{Create a SpatialPolygonsDataFrame or a SpatialLinesDataFrame from the
Stewart Raster}
\usage{
contourStewart(x, breaks, type = "line")
}
\arguments{
\item{x}{object of class raster outputed by the rasterStewart function}

\item{breaks}{NUMERIC, a vector of break values.}

\item{type}{'poly' or 'line'. WARNING: the poly option is experimental and
can output errors}
}
\description{
This function create a SpatialPolygonsDataFrame or a
SpatialLinesDataFrame from the stewart raster
}
\details{
The ouput of the function is a SpatialPolygonsDataFrame or a
SpatialLinesDataFrame
}
\examples{
# Calculate potentials
data(spatData)
}

