% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{CreateDistMatrix}
\alias{CreateDistMatrix}
\title{Create a Distance Matrix Between Two SpatialPointsDataFrame Objects}
\usage{
CreateDistMatrix(knownpts, unknownpts, longlat = FALSE, bypassctrl = FALSE)
}
\arguments{
\item{knownpts}{SP OBJECT (SpatialPointsDataFrame).}

\item{unknownpts}{SP OBJECT (SpatialPointsDataFrame).}

\item{longlat}{LOGICAL, Euclidean distance (FALSE, default) or Great Circle distance (TRUE)}

\item{bypassctrl}{LOGICAL, bypass the distance matrix size control (see Details).}
}
\description{
This function creates a distance matrix between two
SpatialPointsDataFrame Objects
}
\details{
The function returns a full matrix of distances in the metric of the
points if longlat=FALSE, or in kilometers if longlat=TRUE. This is a wrapper
for the \code{\link{spDists}} function.

If the matrix to compute is too large (more than 100,000,000 cells or more than 10,000,000 origins or destinations)
the function sends a confirmation message to warn users about the amount of RAM mobilized.
bypassctrl=TRUE skip this control.
}
\examples{
# Create a SpatialPointsDataFrame grid of spatMask extent and 200 meters
# resolution
data(spatData)
mygrid <- CreateGrid(w = spatMask, resolution = 200)
# Create a distance matrix between known points (spatPts) and mygrid
mymat <- CreateDistMatrix(knownpts = spatPts, unknownpts = mygrid,
                          longlat = FALSE)
mymat[1:5,1:5]
dim(mymat)
}

