% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{CreateGrid}
\alias{CreateGrid}
\title{Create a Regular SpatialPointsDataFrame}
\usage{
CreateGrid(w, resolution)
}
\arguments{
\item{w}{SP OBJECT, (SpatialPointsDataFrame or
SpatialPolygonsDataFrame). The spatial extent of this object is used to
create the regular SpatialPointsDataFrame}

\item{resolution}{INTEGER, resolution of the output grid (in map units).}
}
\description{
This function creates a regular grid of SpatialPointDataFrame
from the extent of a given Spatial*DataFrame and a given resolution.
}
\details{
The output of the function is a SpatialPointsDataFrame of regularly
spaced points with the same extent as \code{w}.
}
\examples{
# Create a SpatialPointsDataFrame grid of spatMask extent and 200 meters
# resolution
data(spatData)
mygrid <- CreateGrid(w = spatMask, resolution = 200)
plot(mygrid, cex = 0.1, pch = ".")
plot(spatMask, border="red", lwd = 2, add = TRUE)
}

