% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{CreatePotentialsContourPoly}
\alias{CreatePotentialsContourPoly}
\title{Create a SpatialPolygonsDataFrame of the Potential Raster Contour Lines}
\usage{
CreatePotentialsContourPoly(potentials.raster, potentials.breaks)
}
\arguments{
\item{potentials.raster}{object of class raster outputed by the
\code{\link{CreateRaster}} function.}

\item{potentials.breaks}{NUMERIC, a vector of break values to map.}
}
\description{
This function creates a SpatialPolygonsDataFrame from the
potential raster outputed by the \code{\link{CreateRaster}}
function.
}
\details{
The ouput of the function is a SpatialPolygonsDataFrame
}
\examples{
# Calculate potentials
data(spatData)
}

