% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{CreateRaster}
\alias{CreateRaster}
\title{Create a Raster from a SpatialPointsDataFrame Potential Layer}
\usage{
CreateRaster(x, mask = NULL)
}
\arguments{
\item{x}{object of class sp (SpatialPointsDataFrame) outputed by the
stewart, huff or reilly function.}

\item{mask}{object of class sp (SpatialPolygonsDataFrame) to clip the map of
potentials.}
}
\description{
This function create a raster layer from SpatialPointsDataFrame
potential layer outputed by the \code{\link{stewart}} function.
}
\details{
The ouput of the function is a raster of the potentials
}
\examples{
# Calculate potentials
data(spatData)
}

