% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoband.R
\name{isobander}
\alias{isobander}
\title{Create Spatial Polygons Contours from a Raster}
\usage{
isobander(x, nclass = 8, breaks = NULL, mask = NULL)
}
\arguments{
\item{x}{sf POINT data.frame; must contain X, Y and OUTPUT fields.}

\item{nclass}{numeric; a number of class.}

\item{breaks}{numeric; a vector of break values.}

\item{mask}{sf POLYGON data.frame; mask used to 
clip contour shapes.}
}
\value{
The output is an sf POLYGON data.frame.
The data frame contains four fields: 
id (id of each polygon), min and max (minimum and maximum breaks of the polygon), 
center (central values of classes).
}
\description{
This function creates spatial polygons of contours from a raster.
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart}, 
\link{quickStewart}, \link{CreateGrid}, \link{CreateDistMatrix}.
}
