% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{PlotPotentialsContour}
\alias{PlotPotentialsContour}
\title{Plot Potentials Contour}
\usage{
PlotPotentialsContour(potentials, potentials.breaks, legend.rnd = 0,
  col.contour = "black")
}
\arguments{
\item{potentials}{object of class sp (SpatialPointsDataFrame or
SpatialPolygonsDataFrame) outputed by the potentials function.}

\item{potentials.breaks}{NUMERIC, a vector of break values to map. If used,
this parameter override \code{typec} and \code{nclass} params}

\item{legend.rnd}{NUMERIC, number of digits used to round the figures
displayed in the legend}

\item{col.contour}{CHARACTER, a color}
}
\description{
This function plot a contour from from SpatialPointsDataFrame
potential layer outputed by the \code{\link{stewart}} function.
}
\details{
This function returns only a contour on the current plot
}
\examples{
# Calculate potentials
data(spatData)
}

