% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{PlotPotentialsRaster}
\alias{PlotPotentialsRaster}
\title{Plot a Potentials Raster}
\usage{
PlotPotentialsRaster(potentials.raster, add = TRUE,
  potentials.breaks = NULL, typec = "equal", nclass = 5, legend.rnd = 0,
  col.raster = colorRampPalette(c("#FEA3A3", "#980000")))
}
\arguments{
\item{potentials.raster}{object of class raster outputed by the
\code{\link{CreateRaster}} function.}

\item{add}{BOOLEAN, classic add parameter}

\item{potentials.breaks}{NUMERIC, a vector of break values to map. If used,
this parameter override \code{typec} and \code{nclass} params}

\item{typec}{CHARACTER, either "equal" or "quantile", how to discretize the
map.}

\item{nclass}{NUMERIC, number of class to map.}

\item{legend.rnd}{NUMERIC, number of digits used to round the figures
displayed in the legend}

\item{col.raster}{FUNCTION, colors to display the map like
\code{\link{colorRampPalette}}.}
}
\description{
This function plot a potential raster outputed by the
\code{\link{CreateRaster}} function.
}
\details{
The ouput of the function is a plotted raster and the (invisible)
list of break values
}
\examples{
# Calculate potentials
data(spatData)
}

