% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{plotReilly}
\alias{plotReilly}
\title{Plot a Reilly Raster}
\usage{
plotReilly(x, add = TRUE, col = rainbow)
}
\arguments{
\item{x}{object of class raster outputed by the
\code{\link{rasterReilly}} function.}

\item{add}{BOOLEAN, classic add parameter}

\item{col}{FUNCTION, colors to display the map, like \code{\link{rainbow}}.}
}
\description{
This function plot a potential raster outputed by the
\code{\link{rasterReilly}} function.
}
\details{
The ouput of the function is a plotted raster.
}
\examples{
# Calculate potentials
data(spatData)
}

