% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickStewart.R
\name{quickStewart}
\alias{quickStewart}
\title{Create a SpatialPolygonsDataFrame of Potentials}
\usage{
quickStewart(spdf, df, spdfid = NULL, dfid = NULL, var, var2 = NULL,
  typefct = "exponential", span, beta, resolution = NULL, mask = NULL,
  breaks = NULL)
}
\arguments{
\item{spdf}{a SpatialPolygonsDataFrame.}

\item{df}{a data frame that contains the values to plot.}

\item{spdfid}{name of the identifier field in spdf, default to the first column 
of the spdf data frame. (optional)}

\item{dfid}{name of the identifier field in df, default to the first column 
of df. (optional)}

\item{var}{name of the numeric field in df used to compute potentials.}

\item{var2}{name of the numeric field in df used to compute potentials, used for ratio (numerator).}

\item{typefct}{character; spatial interaction function. Options are "pareto" 
(means power law) or "exponential".
If "pareto" the interaction is defined as: (1 + alpha * mDistance) ^ (-beta).
If "exponential" the interaction is defined as: 
exp(- alpha * mDistance ^ beta).
The alpha parameter is computed from parameters given by the user 
(\code{beta} and \code{span}).}

\item{span}{numeric; distance where the density of probability of the spatial 
interaction function equals 0.5.}

\item{beta}{numeric; impedance factor for the spatial interaction function.}

\item{resolution}{numeric; resolution of the output SpatialPointsDataFrame
(in map units).}

\item{mask}{SpatialPolygonsDataFrame; mask used to clip contour shapes.}

\item{breaks}{numeric; a vector of break values.}
}
\value{
The ouput of the function is a SpatialPolygonsDataFrame.
}
\description{
This function creates a SpatialPolygonsDataFrame of potential values. 
It allows also to compute directly the ratio between two variables.
}
\details{
This function is mainly a wrapper around stewart, rasterStewart and contourStewart.
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart}, \link{contourStewart}, \link{CreateGrid}, \link{CreateDistMatrix}.
}

