% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterReilly}
\alias{rasterReilly}
\title{Create a Raster from a Reilly SpatialPointsDataFrame Layer}
\usage{
rasterReilly(x, mask = NULL)
}
\arguments{
\item{x,}{SP OBJECT (SpatialPointsDataFrame), output of the \code{reilly} function.}

\item{mask,}{SP OBJECT (SpatialPolygonsDataFrame) to clip the raster with.}
}
\description{
This function creates a raster layer from SpatialPointsDataFrame
potential layer, output of the \code{\link{reilly}} function.
}
\details{
Raster layer with output values (catchment areas).
The raster uses a RAT (\code{\link{ratify}}) that contains the
correspondance between raster values and catchement areas values. Use \code{
unique(levels(rasterName)[[1]])} to see the correpondance table.
}
\examples{
# Calculate potentials
data(spatData)
}

