% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterStewart}
\alias{rasterStewart}
\title{Create a Raster from a Stewart SpatialPointsDataFrame Layer}
\usage{
rasterStewart(x, mask = NULL)
}
\arguments{
\item{x}{SP OBJECT (SpatialPointsDataFrame), output of the \code{stewart} function.}

\item{mask}{SP OBJECT (SpatialPolygonsDataFrame) to clip the raster with.}
}
\description{
This function creates a raster layer from SpatialPointsDataFrame
potential layer, output of the \code{\link{stewart}} function.
}
\details{
Raster layer with output values (potentials).
}
\examples{
# Compute Stewart potentials from known points (spatPts) on a
# grid defined by its resolution
mystewart <- stewart(knownpts = spatPts, varname = "POPULATION",
                     typefct = "exponential", span = 1250, beta = 3,
                     resolution = 200, longlat = FALSE, mask = spatMask)
# Create a raster of potentials values
mystewartraster <- rasterStewart(x = mystewart, mask = spatMask)
plot(mystewartraster)
}

