% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterToContourPoly.R
\name{rasterToContourPoly}
\alias{rasterToContourPoly}
\title{Create Spatial Polygons Contours from a Raster}
\usage{
rasterToContourPoly(r, nclass = 8, breaks = NULL, mask = NULL,
  sf = FALSE)
}
\arguments{
\item{r}{raster; the raster must contain only positive values.}

\item{nclass}{numeric; a number of class.}

\item{breaks}{numeric; a vector of break values.}

\item{mask}{SpatialPolygonsDataFrame or sf POLYGON data.frame; mask used to 
clip contour shapes. 
The mask should have a smaller extent than r.}

\item{sf}{boolean; if TRUE, the output is an sf POLYGON data.frame}
}
\value{
If sf is FALSE, the ouput of the function is a SpatialPolygonsDataFrame. 
If sf is TRUE the output is an sf POLYGON data.frame.
The data frame contains four fields: 
id (id of each polygon), min and max (minimum and maximum breaks of the polygon), 
center (central values of classes).
}
\description{
This function creates spatial polygons of contours from a raster.
}
\examples{
data("spatData")
\dontrun{
mystewart <- stewart(knownpts = spatPts, varname = "Capacite",
                     typefct = "exponential", span = 1000, beta = 3,
                     resolution = 50, mask = spatMask)
# Create a raster of potentials values
mystewartraster <- rasterStewart(x = mystewart)
# Create contour SpatialLinesDataFrame
contourpoly <- rasterToContourPoly(r = mystewartraster,
                                   nclass = 6,
                                   mask = spatMask)
# Created breaks
bks <- sort(unique(c(contourpoly$min, contourpoly$max)))
# Display the map
library(cartography)
opar <- par(mar = c(0,0,1.2,0))
choroLayer(spdf = contourpoly,
           df = contourpoly@data,
           var = "center", legend.pos = "topleft",
           breaks = bks, border = "grey90",
           lwd = 0.2,
           legend.title.txt = "Potential number\\nof beds in the\\nneighbourhood",
           legend.values.rnd = 0)
plot(spatMask, add = TRUE)
propSymbolsLayer(spdf = spatPts, df = spatPts@data, var = "Capacite",
                 legend.title.txt = "Number of beds",
                 col = "#ff000020")
layoutLayer(title = "Global Accessibility to Public Hospitals",
            south = TRUE, sources = "", author = "")
par(opar)
}
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart}, 
\link{quickStewart}, \link{CreateGrid}, \link{CreateDistMatrix}.
}
