% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{reilly}
\alias{reilly}
\title{Calculate reilly}
\usage{
reilly(knownpts, unknownpts = NULL, matdist = NULL, varname,
  typefct = "exponential", span, beta, resolution = 2000, longlat = FALSE,
  mask = NULL)
}
\arguments{
\item{knownpts}{object of class sp (SpatialPointsDataFrame or
SpatialPolygonsDataFrame). It contains the known points to calculate
potentials from.}

\item{unknownpts}{object of class sp (SpatialPointsDataFrame or
SpatialPolygonsDataFrame). It contains points to calculate the potentials on.
When used \code{resolution} is not used.}

\item{matdist}{NUMERIC MATRIX, a distance matrix. Row names match the first
column of the knownpts object dataframe. Column names match the first column
of the knownpts object dataframe. if used \code{typedist} is not used}

\item{varname}{CHARACTER, name of the variable (in the dataframe of the
Spatial*DataFrame) from which potentials are computed.
Quantitative variable with no negative values.}

\item{typefct}{CHARACTER, spatial interaction function. Options are "pareto"
(default) or "exponential".
If "pareto" the interaction is defined as: (1 + alpha * mDistance) ^ (-beta).
If "exponential" the interaction is defined as:
exp(- alpha * mDistance ^ beta).
The alpha parameter is computed from parameters given by the user
(beta and reach).}

\item{span}{NUMERIC, distance where the density of probability of the spatial
interaction function equals 0.5.}

\item{beta}{NUMERIC, impedance factor for the spatial interaction function.}

\item{resolution}{INTEGER, resolution of the output grid (in map units).}

\item{longlat}{BOOLEAN, type of distance. If TRUE longitudz and latitude are
expected in decimal degrees, in WGS84 reference system.}

\item{mask}{object of class sp (SpatialPolygonsDataFrame) to clip the map of
potentials.}
}
\description{
This function compute the potentials of spatial interaction as
defined by J.Q. Stewart (1950).
}
\details{
The ouput value returned by the function is a SpatialPointsDataFrame
of points with the computed potential in the "POTENTIALS" variable
}
\examples{
# Calculate potentials
data(spatData)
}
\references{
Stewart J. Q., Demographic gravitation: evidence and applications
, Sociometry, 11(1-2), 31-58.
}

