% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\docType{package}
\name{SpatialPosition}
\alias{SpatialPosition}
\alias{SpatialPosition-package}
\title{Spatial Position Package}
\description{
Computes spatial position models: Stewart
potentials, Reilly catchment areas, Huff catchment areas.
}

