% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\docType{data}
\name{spatMask}
\alias{spatMask}
\title{SpatialUnits}
\description{
Paris arrondissements and its surrounding communes (SpatialPolygonsDataFrame). Total population in the POPULATION field.
}

