% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\docType{data}
\name{spatUnits}
\alias{spatUnits}
\title{Spatial Units of Paris}
\description{
20 spatial units representing the Paris' districts (*arrondissements*), (SpatialPolygonsDataFrame).
}

