% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{stewartParallel}
\alias{stewartParallel}
\title{Stewart Potentials Parallel}
\usage{
stewartParallel(knownpts, unknownpts = NULL, varname,
  typefct = "exponential", span, beta, resolution = NULL, mask = NULL,
  cl = 2, chunks = 100)
}
\arguments{
\item{knownpts}{sp object (SpatialPointsDataFrame or SpatialPolygonsDataFrame);
this is the set of known observations to estimate the potentials from.}

\item{unknownpts}{sp object (SpatialPointsDataFrame or SpatialPolygonsDataFrame); 
this is the set of unknown units for which the function computes the estimates. 
Not used when \code{resolution} is set up. (optional)}

\item{varname}{character; name of the variable in the \code{knownpts} dataframe 
from which potentials are computed. Quantitative variable with no negative values.}

\item{typefct}{character; spatial interaction function. Options are "pareto" 
(means power law) or "exponential".
If "pareto" the interaction is defined as: (1 + alpha * mDistance) ^ (-beta).
If "exponential" the interaction is defined as: 
exp(- alpha * mDistance ^ beta).
The alpha parameter is computed from parameters given by the user 
(\code{beta} and \code{span}).}

\item{span}{numeric; distance where the density of probability of the spatial 
interaction function equals 0.5.}

\item{beta}{numeric; impedance factor for the spatial interaction function.}

\item{resolution}{numeric; resolution of the output SpatialPointsDataFrame
(in map units). If resolution is not set, the grid will contain around 7250 
points. (optional)}

\item{mask}{sp object; the spatial extent of this object is used to 
create the regularly spaced SpatialPointsDataFrame output. (optional)}

\item{cl}{number of clusters}

\item{chunks}{chunks size of unknowpts}
}
\value{
SpatialPointsDataFrame with the computed potentials in a new field 
named \code{OUTPUT}
}
\description{
This function computes the potentials with parallel computing. 
EXPERIMENTAL use. Only Great Circle Distances are used.
}
\examples{
\dontrun{
if(require(cartography)){
system.time(
  s1 <- stewart(knownpts = nuts3.spdf,resolution = 40000, 
               varname = "pop2008",
               typefct = "exponential", span = 100000,
               beta = 3, mask = nuts3.spdf)
)

system.time(
  s2 <- stewartParallel(knownpts = nuts3.spdf, resolution = 40000, 
                        varname = "pop2008",
                        typefct = "exponential", span = 100000,
                        beta = 3, mask = nuts3.spdf, cl = 4, chunks = 500)
)

identical(s1, s2)

r2 <- rasterStewart(s2)
c2 <- rasterToContourPoly(r2, breaks = c(0,10000,100000,1000000,2004000,
                                         5000000,10040000, 20000000200000000), 
                          mask = nuts3.spdf)

 par(mar = c(0,0,0,0))
 bks <- sort(unique(c(c2$min, c2$max)))
 choroLayer(spdf = c2,
            var = "center", legend.pos = "topleft",
            breaks = bks, border = NA)
}
}
}
\references{
STEWART J.Q. (1942) "Measure of the influence of a population at a distance", Sociometry, 5(1): 63-71.
}
\seealso{
\link{rasterStewart}, \link{plotStewart}, \link{quickStewart},
\link{rasterToContourPoly}, \link{CreateGrid}, \link{CreateDistMatrix}.
}

