% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourStewart.R
\name{contourStewart}
\alias{contourStewart}
\title{Create a SpatialPolygonsDataFrame or a SpatialLinesDataFrame from a 
Stewart Raster}
\usage{
contourStewart(x, breaks, mask, type = "line")
}
\arguments{
\item{x}{raster; output of the \code{\link{rasterStewart}} function.}

\item{breaks}{numeric; a vector of break values.}

\item{mask}{SpatialPolygonsDataFrame; mask used to clip contour shapes.}

\item{type}{character; "poly" or "line". WARNING: the poly option is experimental (see details). It needs the rgeos package.}
}
\value{
The ouput of the function is a SpatialPolygonsDataFrame (\code{type = "poly"}) or a SpatialLinesDataFrame (\code{type = "line"}).
}
\description{
This function create a SpatialPolygonsDataFrame or SpatialLinesDataFrame contour from the Stewart raster.
}
\details{
The steps to obtain a SpatialPolygonsDataFrame: \itemize{
\item{Step 1: Set a resolution in stewart (do not enter an unknownpts layer) and enter a SpatialPolygonsDataFrame as mask.}
\item{Step 2: Do not set a mask in rasterStewart}
\item{Step 3: Enter the step 1 mask SpatialPolygonsDataFrame as mask in contourStewart}
}
}
\examples{
# type = "poly"
mystewart <- stewart(knownpts = spatPts, varname = "Capacite",
                     typefct = "exponential", span = 1000, beta = 3,
                     resolution = 50, longlat = FALSE,
                     mask = spatMask)
# Create a raster of potentials values
mystewartraster <- rasterStewart(x = mystewart)
# Display the raster and get break values
break.values <- plotStewart(x = mystewartraster)
# Create contour SpatialLinesDataFrame
mystewartcontourpoly <- contourStewart(x = mystewartraster,
                                       breaks = break.values,
                                       mask = spatMask,
                                       type = "poly")

plot(spatMask, col = "grey80", border = "grey50")
plot(mystewartcontourpoly, col = "#0000ff50", border = "grey40",
     add = TRUE)
plot(spatPts, cex = 0.5, pch = 20, col  = "#ff000050", add = TRUE)
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart}, \link{contourStewart}, \link{CreateGrid}, \link{CreateDistMatrix}.
}

