% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Potentials.R
\name{CreateDistMatrix}
\alias{CreateDistMatrix}
\title{Create a Distance Matrix Between Two SpatialPointsDataFrame Objects}
\usage{
CreateDistMatrix(knownpts, unknownpts, longlat = FALSE)
}
\arguments{
\item{knownpts}{object of class SpatialPointsDataFrame.}

\item{unknownpts}{object of class SpatialPointsDataFrame.}

\item{longlat}{if FALSE, Euclidean distance, if TRUE Great Circle distance}
}
\description{
This function creates a distace matrix between two
SpatialPointsDataFrame Objects
}
\details{
The function returns a full matrix of distances in the metric of the
 points if longlat=FALSE, or in kilometers if longlat=TRUE. This is a wrapper
  around the \code{\link{spDists}} function.
}
\examples{
# Calculate potentials
data(spatData)
}

