% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{CreateDistMatrix}
\alias{CreateDistMatrix}
\title{Create a Distance Matrix Between Two SpatialPointsDataFrame Objects}
\usage{
CreateDistMatrix(knownpts, unknownpts, longlat = FALSE)
}
\arguments{
\item{knownpts}{SP OBJECT (SpatialPointsDataFrame).}

\item{unknownpts}{SP OBJECT (SpatialPointsDataFrame).}

\item{longlat}{LOGICAL, Euclidean distance (FALSE, default) or Great Circle distance (TRUE)}
}
\description{
This function creates a distace matrix between two
SpatialPointsDataFrame Objects
}
\details{
The function returns a full matrix of distances in the metric of the
 points if longlat=FALSE, or in kilometers if longlat=TRUE. This is a wrapper
  for the \code{\link{spDists}} function.
}
\examples{
# Create a SpatialPointsDataFrame grid of spatMask extent and 200 meters
# resolution
data(spatData)
mygrid <- CreateGrid(w = spatMask, resolution = 200)
# Create a distance matrix between known points (spatPts) and mygrid
mymat <- CreateDistMatrix(knownpts = spatPts, unknownpts = mygrid,
                          longlat = FALSE)
mymat[1:5,1:5]
dim(mymat)
}

