% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{CreateGrid}
\alias{CreateGrid}
\title{Create a Regular SpatialPointsDataFrame}
\usage{
CreateGrid(w, resolution)
}
\arguments{
\item{w}{object of class sp (SpatialPointsDataFrame or
SpatialPolygonsDataFrame). The spatial extent of this object is used to
create the regular SpatialPointsDataFrame}

\item{resolution}{INTEGER, resolution of the output grid (in map units).}
}
\description{
This function creates a regular grid of SpatialPointDataFrame
from the extent of a given Spatial*DataFrame and a given resolution.
}
\details{
The output of the function is a SpatialPointsDataFrame of regularly
spaced points with the same extent as \code{w}.
}
\examples{
# Calculate potentials
data(spatData)
}

