% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridandmat.R
\name{CreateGrid}
\alias{CreateGrid}
\title{Create a Regularly Spaced Points Grid}
\usage{
CreateGrid(w, resolution)
}
\arguments{
\item{w}{sp or sf object; the spatial extent of this object is used to 
create the regular grid.}

\item{resolution}{numeric; resolution of the grid (in map units). If 
resolution is not set, the grid will contain around 7500 points. (optional)}
}
\value{
The output of the function is a regularly spaced points grid with the 
extent of \code{w}. If \code{w} is an sp object, the output is a 
SpatialPointsDataFrame; if \code{w} is an sf object, the output is an sf 
POINT data.frame.
}
\description{
This function creates a regular grid of points 
from the extent of a given spatial object and a given resolution.
}
\examples{
# Create a SpatialPointsDataFrame grid of spatMask extent and 200 meters 
# resolution
library(sp)
data(spatData)
mygrid <- CreateGrid(w = spatMask, resolution = 200)
plot(mygrid, cex = 0.1, pch = ".")
plot(spatMask, border="red", lwd = 2, add = TRUE)
}
\seealso{
\link{CreateDistMatrix}
}
