% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{plotHuff}
\alias{plotHuff}
\title{Plot a Huff Raster}
\usage{
plotHuff(x, add = TRUE, breaks = NULL, typec = "equal", nclass = 5,
  legend.rnd = 0, col = colorRampPalette(c("#E8F6A4", "#445200")))
}
\arguments{
\item{x}{object of class raster outputed by the
\code{\link{rasterHuff}} function.}

\item{add}{BOOLEAN, classic add parameter}

\item{breaks}{NUMERIC, a vector of break values to map. If used,
this parameter override \code{typec} and \code{nclass} params}

\item{typec}{CHARACTER, either "equal" or "quantile", how to discretize the
map.}

\item{nclass}{NUMERIC, number of class to map.}

\item{legend.rnd}{NUMERIC, number of digits used to round the figures
displayed in the legend}

\item{col}{FUNCTION, colors to display the map like
\code{\link{colorRampPalette}}.}
}
\description{
This function plot a potential raster outputed by the
\code{\link{rasterHuff}} function.
}
\details{
The ouput of the function is a plotted raster and the (invisible)
list of break values
}
\examples{
# Calculate potentials
data(spatData)
}

