% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{plotReilly}
\alias{plotReilly}
\title{Plot a Reilly Raster}
\usage{
plotReilly(x, add = FALSE, col = rainbow)
}
\arguments{
\item{x}{RASTER OBJECT, output of the \code{\link{rasterStewart}} function.}

\item{add}{LOGICAL, add parameter for the \code{plot} function.}

\item{col}{FUNCTION, color ramp produced by functions such as \code{\link{colorRampPalette}}.}
}
\description{
This function plots the raster produced by the \code{\link{rasterReilly}} function.
}
\details{
Display the raster nicely.
}
\examples{
data(spatData)
row.names(spatPts) <- spatPts$CodHop
# Compute Reilly catchment areas from known points (spatPts) on a
# grid defined by its resolution
myreilly <- reilly(knownpts = spatPts, varname = "Capacite",
               typefct = "exponential", span = 1500, beta = 3,
               resolution = 200, longlat = FALSE, mask = spatMask)
# Create a raster of reilly values
myreillyraster <- rasterReilly(x = myreilly, mask = spatMask)
# Plot the raster nicely
plotReilly(x = myreillyraster)
}

