% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{plotReilly}
\alias{plotReilly}
\title{Plot a Reilly Raster}
\usage{
plotReilly(x, add = TRUE, col = rainbow)
}
\arguments{
\item{x,}{RASTER OBJECT, output of the \code{\link{rasterStewart}} function.}

\item{add,}{LOGICAL, add parameter for the \code{plot} function.}

\item{col,}{FUNCTION, color ramp produced by functions such as \code{\link{colorRampPalette}}.}
}
\description{
This function plots the raster produced by the \code{\link{rasterReilly}} function.
}
\details{
Print the raster.
}
\examples{
# Calculate potentials
data(spatData)
}

