% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterHuff}
\alias{rasterHuff}
\title{Create a Raster from a Huff SpatialPointsDataFrame Layer}
\usage{
rasterHuff(x, mask = NULL)
}
\arguments{
\item{x}{object of class sp (SpatialPointsDataFrame) outputed by the
huff function.}

\item{mask}{object of class sp (SpatialPolygonsDataFrame) to clip the map of
huff catchement areas.}
}
\description{
This function create a raster layer from gridded
SpatialPointsDataFrame layer outputed by the \code{\link{huff}}.
}
\details{
The ouput of the function is a raster of the Huff values.
}
\examples{
# Calculate potentials
data(spatData)
}

