% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterHuff}
\alias{rasterHuff}
\title{Create a Raster from a Huff SpatialPointsDataFrame Layer}
\usage{
rasterHuff(x, mask = NULL)
}
\arguments{
\item{x}{SP OBJECT (SpatialPointsDataFrame), output of the \code{huff} function.}

\item{mask}{SP OBJECT (SpatialPolygonsDataFrame) to clip the raster with.}
}
\description{
This function creates a raster layer from SpatialPointsDataFrame
potential layer, output of the \code{\link{huff}} function.
}
\details{
Raster layer with output values (catchment areas).
}
\examples{
data(spatData)
# Compute Huff catchment areas from known points (spatPts) on a
# grid defined by its resolution
myhuff <- huff(knownpts = spatPts, varname = "POPULATION",
                     typefct = "exponential", span = 1250, beta = 3,
                     resolution = 200, longlat = FALSE, mask = spatMask)
# Create a raster of huff values
myhuffraster <- rasterHuff(x = myhuff, mask = spatMask)
plot(myhuffraster)
}

