% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterReilly}
\alias{rasterReilly}
\title{Create a Raster from a Reilly SpatialPointsDataFrame Layer}
\usage{
rasterReilly(x, mask = NULL)
}
\arguments{
\item{x,}{SP OBJECT (SpatialPointsDataFrame), output of the \code{reilly} function.}

\item{mask,}{SP OBJECT (SpatialPolygonsDataFrame) to clip the raster with.}
}
\description{
This function creates a raster layer from SpatialPointsDataFrame
potential layer, output of the \code{\link{reilly}} function.
}
\details{
Raster layer with output values (catchment areas).
The raster uses a RAT (\code{\link{ratify}}) that contains the
correspondance between raster values and catchement areas values. Use \code{
unique(levels(rasterName)[[1]])} to see the correpondance table.
}
\examples{
data(spatData)
row.names(spatPts) <- spatPts$CodHop
# Compute Reilly catchment areas from known points (spatPts) on a
# grid defined by its resolution
myreilly <- reilly(knownpts = spatPts, varname = "Capacite",
               typefct = "exponential", span = 1250, beta = 3,
               resolution = 200, longlat = FALSE, mask = spatMask)
# Create a raster of reilly values
myreillyraster <- rasterReilly(x = myreilly, mask = spatMask)
plot(myreillyraster)
# Correspondance between raster values and reilly areas
head(unique(levels(myreillyraster)[[1]]))
}

