% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterReilly}
\alias{rasterReilly}
\title{Create a Raster from a Reilly SpatialPointsDataFrame Layer}
\usage{
rasterReilly(x, mask = NULL)
}
\arguments{
\item{x}{object of class sp (SpatialPointsDataFrame) outputed by the
huff function.}

\item{mask}{object of class sp (SpatialPolygonsDataFrame) to clip the map of
reilly catchement areas.}
}
\description{
This function create a raster layer from gridded
SpatialPointsDataFrame layer outputed by the \code{\link{reilly}} function.
}
\details{
The ouput of the function is a raster of the Reilly values.
The raster uses a RAT (\code{\link{ratify}}) that contains the
correspondance between raster values and catchement areas values. Use \code{
unique(levels(rasterName)[[1]])} to see the correpondance table.
}
\examples{
# Calculate potentials
data(spatData)
}

