% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stewarthuffreilly.R
\name{rasterStewart}
\alias{rasterStewart}
\title{Create a Raster from a Stewart SpatialPointsDataFrame Layer}
\usage{
rasterStewart(x, mask = NULL)
}
\arguments{
\item{x}{object of class sp (SpatialPointsDataFrame) outputed by the
stewart function.}

\item{mask}{object of class sp (SpatialPolygonsDataFrame) to clip the map of
potentials.}
}
\description{
This function create a raster layer from gridded
SpatialPointsDataFrame layer outputed by the \code{\link{stewart}}.
}
\details{
The ouput of the function is a raster of the Stewart potentials
}
\examples{
# Calculate potentials
data(spatData)
}

