% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SpatialPosition}
\alias{SpatialPosition}
\alias{SpatialPosition-package}
\title{Spatial Position Package}
\description{
Computes spatial position models: \itemize{
\item{Stewart potentials,}
\item{Reilly catchment areas,} 
\item{Huff catchment areas.}
}
An introduction to the package conceptual background and usage 
(see \code{vignette(topic = "SpatialPosition")}) an a Stewart potentials 
use case (see \code{vignette(topic = "StewartExample")}) are proposed in vignettes.
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart}, 
\link{contourStewart}, \link{huff}, \link{rasterHuff}, \link{plotHuff},\link{reilly}, 
\link{rasterReilly}, \link{plotReilly}, 
\link{CreateGrid}, \link{CreateDistMatrix}.
}

