% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SpatialPosition}
\alias{SpatialPosition}
\alias{SpatialPosition-package}
\title{Spatial Position Package}
\description{
Computes spatial position models: \itemize{
\item{Stewart potentials,}
\item{Reilly catchment areas,} 
\item{Huff catchment areas.}
}
An introduction to the package conceptual background and usage: \cr
 - \code{vignette(topic = "SpatialPosition")}\cr
A Stewart potentials use case:\cr
 - \code{vignette(topic = "StewartExample")}.
}
\references{
COMMENGES H., GIRAUD, T., LAMBERT, N. (2016) "ESPON FIT: Functional Indicators for Spatial-Aware Policy-Making", 
Cartographica: The International Journal for Geographic Information and Geovisualization, 51(3): 127-136.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/riatelab/SpatialPosition}
  \item Report bugs at \url{https://github.com/riatelab/SpatialPosition/issues}
}

}
\author{
\strong{Maintainer}: Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID})

Authors:
\itemize{
  \item Hadrien Commenges
}

Other contributors:
\itemize{
  \item Joël Boulier [contributor]
}

}
