% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SpatialPosition}
\alias{SpatialPosition}
\alias{SpatialPosition-package}
\title{Spatial Position Package}
\description{
Computes spatial position models: Stewart
potentials, Reilly catchment areas, Huff catchment areas.
}
\seealso{
\link{stewart}, \link{rasterStewart}, \link{plotStewart},
\link{contourStewart}, \link{huff}, \link{rasterHuff}, \link{plotHuff},\link{reilly},
\link{rasterReilly}, \link{plotReilly},
\link{CreateGrid}, \link{CreateDistMatrix}.
}

