% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SpatialPosition}
\alias{SpatialPosition}
\title{Spatial Position Package}
\description{
Computes spatial position models: \itemize{
\item{Stewart potentials,}
\item{Reilly catchment areas,} 
\item{Huff catchment areas.}
}
An introduction to the package conceptual background and usage: \cr
 - \code{vignette(topic = "SpatialPosition")}\cr
A Stewart potentials use case:\cr
 - \code{vignette(topic = "StewartExample")}.
}
\references{
COMMENGES H., GIRAUD, T., LAMBERT, N. (2016) "ESPON FIT: Functional Indicators for Spatial-Aware Policy-Making", 
Cartographica: The International Journal for Geographic Information and Geovisualization, 51(3): 127-136.
}
